refactor(*): start extraction from ClausePlugin
